+++ src/Configurations/src/Configs/ArsenalConfig/WeaponBehaviors/TimeExpirable.java
@@ -19,7 +19,7 @@ public class TimeExpirable extends WeaponBehavior{
     }
 
     @Override
-    public void update(long ms) {
+    public void update(double ms) {
         //TODO
     }
