+++ src/Configurations/src/Configs/ShooterConfig/Shooter.java
@@ -57,7 +57,7 @@ public class Shooter implements Updatable , Configurable {
     }
 
     @Override
-    public void update(long ms) {
+    public void update(double ms) {
         //only shooting radially rn
         if(ms%getRateOfFire()==0) {
             ActiveLevel myActiveLevel =  getMyShootable().getWeaponConfig().getMyArsenal().getGame().getActiveLevel();
